Fixed bug with long conditions.

+++ simulator_engine/src/nl/tue/bpmn/parser/ConditionEvaluator.java
@@ -105,8 +105,8 @@ public class ConditionEvaluator {
 		if (dataItemValue == null){
 			return false;
 		}
-		long numericValue = Long.parseLong(dataItemValue);
-		long compareValue = Long.parseLong(condition.NUMBER().getText());
+		Double numericValue = Double.parseDouble(dataItemValue);
+		Double compareValue = Double.parseDouble(condition.NUMBER().getText());
 		String comparator = condition.COMPARATOR().getText();
 		if (comparator.equals("=")){
 			return numericValue == compareValue;
